Add static site
I haven't really done anything with shiftingpixel.com in a long time, so
it no longer makes sense to host it as a dynamic website. I disabled
comments, swapped out the search box for a Google site search, and
removed some RSS links before grabbing the static assets using:

  wget -mk http://shiftingpixel.com

I noticed that the `-k` flag made most of the links to the individual
posts not as nice looking as it had previously been, but without the
flag a bunch of stuff is broken. I don't really care too terribly much
at this point as long as the old URLs still work, which they do.